Merged PR 568516: Change and  binding redirects to avoid failures running tests from IDE
This change extracts binding redirects information into a helper method in `BuildXLSdk` and it also adds special comments into the extracted helper and into `config.dsc`. This should help to avoid regressions in the future.

The change also fixes a NRE in tests that was caused by an attempt to reuse not fully initialized `LocalRedisProcessDatabase` instance.