Merged PR 568305: Pin dedup content during incorporate
Backed dedup content is not supported by BuildCache. As a workaround, pin all content in the client when calling incorporate.

It was necessary to add a pin method in the dedup content session which also takes in a ttl, since we want to make sure that it matches the ttl of the fingerprint

Related work items: #1749046