Merged PR 569349: Expose undeclared reads + writes on the same path as a read violation
This change inverts the perspective for the DFA that ocurrs when a pip reads a file without declaration (possible when undeclared source reads mode is on) that is also written to by another pip. This used to be presented as a write violation to the user (being the writer the culprit), but the most likely case is the scenario where a declared dependency between both pips is missing, so a read violation, where the reader is to blame, is more natural since it is the reader where the dependency needs to be declared.

This is just changing the message we give to the user. No fundamental changes involved.

Related work items: #1748568