Merged PR 570456: Fix output directory exclusions by properly configuring the pip manifest
Output directory exclusions was working by an unfortunate combination of events:
* Accesses under exclusions were not added to opaques (which was fine)
* Those skipped writes were misclassified by the observed input processor as reads
* The observed input processor would then report those as violations

So everything seemed to work. But using exclusions in combination with undeclared source reads surfaced the problem, where violations were not generated.

The fix here is to add the exclusions directly into the pip manifest. This also simplifies things a little bit and we don't need any special treatment in the sandboxed pip executor for exclusions anymore. This is probably what I should have done to begin with.