Make TTL of file content table entries settable from command line argument
This change also increases the upper bound of TTL to `ushort`, but the default TTL is still set to `byte.MaxValue`.

Office has been experiencing FCT trashing because the TTL is too small as they are now sharing FCT over multiple enlistments. Also, consider that Office developers do obuild a lot of times, and so when they do ohome afterward, they are losing a lot of entries in FCT.