Merged PR 571084: Change grpcPort to uint in ContentStoreApp.Service
When trying to run ContentStoreApp, I got the error:
`Object of type 'System.UInt32' cannot be converted to type 'System.Int32'.`

This was because of the fact that the default value for `GrpcDisabledPort` in the `service` verb for CLAP is a `uint`, while the property type was an `int`. Changing the property type to `uint` fixes the issue.