Merged PR 570533: Don't trace skipped proactive copies
We want to be able to do more analysis on proactive copy retries, but the one thing that is stopping us is that we currently don't trace all successful copies because we fear that it will introduce too much trace volume.

In this case, I want to be able to not worry about tracing copies that we skipped because there are already enough replicas available.

Once this change is in, we should be able to enable `TraceProactiveCopies` everywhere.