Merged PR 570711: Set shorter fingerprint store GC cancelation timer for short builds
1. Set the cancellation timeout to 1/3 of the fingerprintstore in use time or 10 secs, whichever is smaller.

2. Reduce the log level of GC cancellation to information.
Reason: It happens in each build to try to garbage collect those out of date entries/columns for 10 seconds. When time's up, GC will be cancelled and the rest could be done in next build. It is not a big deal to defer the rest of the collection later.

Related work items: #1756920