Merged PR 571695: Refine supersession tracking in file change tracker
Let P be a pip that has output directory D. After execution, P will have some nested directory created underneath D, say D\E.

1. Execute P, expect cache miss.
2. Modify P's input, and execute P again, expect cache miss.
3. Execute P, expect not scheduled, but it is scheduled.

The reason why P gets dirty in the 3rd build is because the deletion of D\E before P is executed in the 2nd build. D\E is a directory and is not superseded.

This change introduces a mode that can be used to supersede all subpaths, instead of file paths. To be on the safe side, the superseding of container path is only used if the change is directory deletion. If the change is rename, then superseding is not used in order to avoid orphaning tracked children.

To be effective, we try to avoid doing renaming as much as possible. This is done by forcing posix delete.

Users can go back to the legacy behavior (if problematic) using `/fileChangeTrackerSupersedeMode:FileOnly`. Or if posix delete is causing problem, then `/posixDeleteMode:RunLast`.

Related work items: #1764535