Merged PR 570862: Exclude replicated locations from inside-ring proactive copy
Proactive copy should never try to push to a machine which already has the content. Currently, we don't exclude these machines from the candidates for the inside-ring copy. This PR changes that.

If after filtering there are no candidates left, it will be considered a failure since we didn't achieve the purpose of replicating the content. Note that the overall proactive copy might still succeed if the outside-ring build is successful.