Merged PR 572256: [JavaScript] Add the ability to untrack a relative path for all projects
It is not uncommon in the JS world that the same cone (same relative to each project in the build) has to be untracked. Junctions are usually the culprit. This PR adds a way to define a relative path as untracked for all projects, so it applies to both the current project and its dependencies.