Merged PR 572311: Use non-blocking materialization concurrency limiter
The old version of this code was using a non-async callback provided to `ActionBlockSlim` (even though the async callback existed).

The issue with the old version was that it was blocking a thread from a thread pool with no good reason. `ActionBlockSlim` uses non-long-running tasks for processing work items, and blocking the task means we're blocking a thread pool thread.

This PR makes the callback provided to `ActionBlockSlim` fully asynchronous to avoid unnecessary block of the worker threads.