Merged PR 572333: Preserving ObservationFlags for AbsentPathProbes
We did not preserve ObservationFlags for AbsentPathProbe. That's why, the probes for absent files were converted to FileContentRead when the files become existent instead of ExistingFileProbe.

For Office, NtCreateFile is called for a file with a trailing backslash. As the access has IsDirectoryLocation flag and the path exists as a file, it becomes AbsentPathProbe after execution. During cache lookup, we have FileContentRead for that ObservedInput because we do not preserve IsDirectoryLocation flag. That's why, that pip was always a miss.

I think that the changes will resolve many cache misses for Cosine builds too because I saw a lot of similar misses in Cosine builds => AbsentPathProbe after execution, FileContentRead after cachelookup for the same path.

The changes also fix a 3-year old, active bug: Bug #1087986

Related work items: #1087986, #1724406