Merged PR 572305: Adding telemetry for get blob from EventHub durations
Currently we already trace the duration for GetDeserializeAndDispatchBlobEventAsync, and we have the time it takes to get the blob from TryGetFileAsync. Now we will separate the time it takes to "Get and Deserialize" from "Dispatch" blob events, and add that information to the trace.

Also adding BlobID to the trace