Enable file access correlation in Detours
We introduce id and correlation id in file operation context so that some operations that have multiple file accesses can be correlated. Those operations include copy file, move/rename file/directory, hard link creation.

This feature is needed by quick build as they try to figure out copy source and destination to allow for some optimization.

The id and correlation id are currently only added to the Detours listener. They can be added to the `ReportedFileAccess` as well, but since there is no use of them in BuildXL case, it is considered out of scope for this work.