Merged PR 573376: [LinuxSandbox] Detect and report dynamically loaded libraries
Implemented using the auditing API provided by the GNU dynamic linker to observe all dynamically loaded shared libraries:

https://man7.org/linux/man-pages/man7/rtld-audit.7.html

Although the same library can be added to both LD_PRELOAD and LD_AUDIT at the same time, I couldn't do that with `libDetours.so` because it calls `dlsym` which is not available in audit callbacks.  That's why the Linux sandbox code is compiled to two (almost identical) libraries, `libDetours.so` and `libBxlAudit.so`; interposing is enabled in the former and disabled in the latter.  Both of them use the same `BxlObserver` code to check and report file accesses as soon as they happen.

Related work items: #1741898, #1750766