Merged PR 572671: Make PerformProactiveCopyAsync include IOGate wait time in operation duration
Currently, `PerformProactiveCopyAsync` duration does not include the IOGate wait time. This is misleading as it can make us think that operations are really quick when in reality they may be taking an extremely long time.

Moving the `GatedOperationAsync` to inside the `PerformOperationAsync` call so that the IOGate wait time is still included.

Several variables had to be moved outside of the `GatedOperationAsync` operation, to make them available for the trace message. Also added `OperationContext.WithTimoutAsync` to have a simple, centralized way of imposing timeouts on operations without any tracing implications.