Merged PR 574012: Fix MaterializeFileCommand serialization problem
We build daemons and then use them in the same build (instead of using daemons from LKG). So it means that daemons will be using serialization logic that might be different from the one used in the build that is currently running (ie., LKG bits).

This PR fixes the out-of-sync serialization between latest master and LKG.

Related work items: #1765145