Merged PR 573259: Support for adding paths to graph file system
Adding /graphFileSystemInput, so devs can pass a list of directories to be added to the graph file system. This was an ask from Cosine, and it can be useful to any customers using SOD. Pips probe include directories (MIDL) that haven't been produced yet. Those directories are produced later during the build as a part of SOD. We add the given directories to the graph file system, so they would exist as a directory throughout the build.

It will significantly improve the cache hit rate for Cosine. Those kind of misses will be fixed:

"Paths": {
          "Changed": {
            "b:\\public\\amd64fre\\onecore\\internal\\base\\inc\\appmodel\\runtime\\winrt": {
              "Old": "AccessType: ExistingDirectoryProbe",
              "New": "AccessType: AbsentPathProbe"
            }
          }

Related work items: #1757265