Merged PR 576000: Add an additional timeout for proactive copy IOGate
Currently we see that there are a lot of cases where the IO gate for proactive operations gets backed up up to days at a time. The intent of this PR is to make sure that there is a limit to how much we wait for the IO gate.

Also, in this PR I decided to move proactive tests to their own file since it was becoming too big. None of the tests were modified.