Merged PR 576706: Allow creation junctions to non-existent targets
Creating a junction to a non-existing target is a valid operation. The target could have been created outside of the build and later removed by the user. But more importantly, the target could have been created by the same pip that creates the junction, but on cache replay, these two operations race. IF we block the creation of a junction when the target is non-existent, cache replay may non-deterministically fail

Related work items: #1771079