Merged PR 574955: Make sure that replaceExisting is honored by FileSystem in macOS when creating hardlinks
There was an issue with a customer using drop, where the following was happening:
1. Drop wants some content, checks the cache, and the file is not there
2. Drop downloads to path A
3. Drop pushes the content to the cache, via a hardlink, which ends up in path B
4. Something goes wrong with the drop operation, retries download.
5. Checks in the cache, and gets a hit.
6. Tries to place with hardlink into location A.
7. File already exists, so hardlink fails with AccessDenied
8. Cache falls back to copy, which calls `FileSystem.OpenAsync` with `FileMode.Create`.
9. FileMode.Create has different behaviors in Unix and Windows
   a. Windows: decouples the file from the hardlink and creates a new file
   b. Unix: Does not decouple the file and truncates
10. Both locations A and B are now truncated to be empty, we do the copy, and the hash is obviously wrong, since it is the empty file now.

This PR makes sure that we delete path A when trying to create the hardlink and the destination already exists, rather than failing with AccessDenied.

Added some unit tests that reproduced the issue. `PlaceFileWithReplaceExistingToHardlinkSource` is exactly the scenario that was failing. With the current changes, the test now passes. The rest of the tests were added to better understand how operations differed between operating systems. For these tests to run, and to make sure we don't regress, I've also removed `restrictTestRunToSomeQualifiers` from the `.dsc` file.

Related work items: #1772905