Merged PR 577056: Add troubleshooting tips for locked output directory error
If the client bxl.exe is externally killed while the server process is shutting down, subsequent builds will hit an error until the server process finishes its shut down. The multiple ctrl-c scenario is fine as the client bxl.exe process will absorb many of them while it waits for the server to shut down cleanly. But parent processes of the bxl.exe client won't necessary pass all ctrl-c instances through without externally killing it.

It isn't clear to BuildXL what the user's intent is when it falls into the scenario of not being able to lock an output directory, so automatically killing concurrently running processes would not be appropriate. We can update the error message to better explain the situation and allow the user to manually kill bxl.exe if that is desired.

Related work items: #1728739