Merged PR 576854: Classify directory creation under opaques as always existent
Consider the case of a directory being created under an opaque, where no files are created underneath. On cache replay, this is always a miss, since the directory creation is not recorded, and therefore the Output filesystem is unaware of it. This change extends an existing codepath that makes dir probes under opaques to be always existent to also include write accesses on the directory (which essentially means directory creation).
@<Semih Okur> , this is the case we discussed. If you think we need to wait longer to be confident about your original change, I'm fine with that, but if you think this change is safe enough, the work on Midgard needs this on master.

Related work items: #1771079