Merged PR 575933: Build Manifest: Generating Sha256 hash in DropDaemon and sending it to XLG
This PR adds a VSO to SHA-256 mapping into cache. The SHA-256 hash is required for Build Manifest generation.
The SHA-256 computation is requested from the Drop Daemon for each file uploaded into Drop.
These values will be sent to the master and added into a BuildManifest.json file in the next PR.
The mapping is currently `1 VSO Hash : 1 SHA-256 Hash`
It can be changed to a group mapping `N vso : N sha` but even a single file change within that group will require materializing the remaining files, which is probably more expensive. An optimization for this would be introducing a new cache API that uses Redis (since this can be considered as a string to string mapping)
Since this feature will not be used anytime soon, this optimization can be picked up in the next quarter.

Related work items: #1763889