Merged PR 577305: More info for non up-to-date ouputs during materialization
More info for non up-to-date ouputs during materialization

We are materializing outputs for the case where no outputs are stored to the cache. At this point, the cache look-up has determined that all output targets are up-to-date (i.e., they have correct hashes). However, upon output materialization, it turns out that the target is no longer up-to-date. (Customers claim that they did not modify the file nor nothing in the build modifies the file.)

The new log and info will show what hashes seen after cache look-up, and the expected hashes of the targets. One should check CopyingPipOutputToLocalStorage log for the actual hashes on disk. The reason for showing the cache look-up content hash is to ensure that the file has been checked to be up-to-date during the cache look-up.

Related work items: #1749016