Merged PR 578272: QTest: Change default RetryMode to Full
Currently Office Motif Test Adapter does not honor the FullyQualifiedName argument passed to VsTest. This causes QTest to fail because in a retry attempt more than expected test cases are run.

This is a temporary fix to revert QTest Retry Mode to the older behavior, and we will work on fixing this for all adapters on QTest side before reenabling the Failed-Only mode.