Merged PR 577995: [JavaScript] Deal with the case of an absent script command in a more intuitive way
If the following dependency is specified prebuild <- build <- postbuild, but in a given project 'build' is missing, postbuild won't depend on anything. A more intuitive treatment is to make postbuild depend on prebuild when build is missing.
This PR makes sure that whenever there is an absent script, its declared dependents get the closes present scripts upstream

Related work items: #1771637