Merged PR 577815: Fix error propagation in TaskUtilities.WhenAllWithCancellation
This PR fixes the issue in redis layer when transient error was causing the whole operation to fail.

The actual issue was in `TaskUtilities.WhanAllWithCancellation`. The method didn't propagate the errors when given tasks were failing. In this case, the callback provided to `RetryStrategy` was finishing successfully, and the call to `NotifyConsumersOfSuccess` was failing with an error, because batch actually failed.

This PR also enables tracient for all the errors in redis layer unconditionally, because having this tracing on all the time is helpful during such kind of investigations and because the errors are not very frequent its ok to have them all the time.

Related work items: #1778090