Merged PR 577563: Add operation limit to redis blob operations and optimize interactions
Make sure that we can only perform a limited amount of operations per Redis instance per machine. The limit has the form of "x operations per y minutes" and is enforced by the new class `OperationCountLimit`.

I opted to do this with a limit count per time period rather than using an IOGate because I really did not want for this to start backing up the same way we've seen proactive copies back up behind an IOGate. Besides, the intended design of having small files in Redis was always to use it as an optimization, and to be able to control how much stress we put into Redis. Perviously, this mean only total memory, but server load has become an issue lately. By controlling the maximum amount of operations, we have a very good handle of this.

A second change done in the PR is to make sure that we don't contact Redis when we know its capacity for blobs has been reached. Previously, when adding a blob, we first checked whether the blob existed or not and if it did, even if capacity had been reached, we returned success. This change will make it so that that particular scenario will now fail, since we value more not overloading Redis than succeeding in that corner case.

Finally, made sure that we can create separate connections for these blob operations in Redis, since we've also seen that sometimes long-running operations can block future incoming operations. This was made behind a flag `UseSeparateConnectionForRedisBlobs`

Related work items: #1772434, #1772641