Merged PR 578948: Introduce FileAndParents supersede mode to address slowness in file change tracker
Last time we introduced a supersede mode where we added a supersession limit for path and its parents, the All supersede mode. In this mode, for the parents, we keep opening a handle and doing point query to get the USN id, which turns out to be extremely expensive.

This new mode, FileAndParents does similar thing, but instead of doing expensive operations, it simply uses the file's USN id as the supersession limit of its parents. This mode is safe because parent paths cannot be modified if handle of its children are being opened. If the parent paths are in different volumes, then the supersession limits will not affect the change, and in that case processing the result of journal scan is the same as FileOnly mode.

This change address the perf. regression in file change tracker when the All supersede mode is in used, but it keeps the goodness of not losing incremental scheduling.

Related work items: #1779640