Merged PR 576170: Move bandwidth checker to GrpcCopyClient and add bandwidth check capability to pushed copies
The main purpose of the PR is to add bandwidth checker capabilities to push copies. However, since there is also a desire to move the checker into the gRPC copier, much more changes were required.

Moved the `BandwidthChecker` into the gRPC copier instead of having a `BandwidthCheckedCopier`. This is desirable because now the checker won't be checking connection time and it'll measure just while streaming content. As with normal copies, different attempts can have different bandwidth requirements. This is achieved by `DistributedContentCopier`. I felt it was important for this logic to live entirely in one place, since this is all an effort to consolidate the behavior pushed and pulled copies.

As an additional quality of life improvement, we will now log the time spent connecting to the server when the copy is successfully attempted, so that we can differentiate between time spent copying and time spent connecting.