Merged PR 579694: Fix invalidoperationexception when accessing moduleWorkerMapping
As ModuleWorkerMapping is updated on another thread, we can rarely get an InvalidOperationException when using Any() in Scheduler. There is no need to make it thread-safe if we use Count. We do not care whether we get a stale value or not. What we care is just the first element.