Merged PR 578202: Build Manifest: XLG to BuildManifest.json on master
In this PR master DropDaemon is calling an API to obtain a BuildManifestData object generated in the ApiServer.
BuildManifestData object contains all the information required for generating the BuildManifest.json file.
In the next PR this file will be created locally (in the drop daemon) and uploaded as a part of the finalize step.

Related work items: #1763898