Merged PR 580692: Make directory fingerprint sensible to new undeclared directories for MinimalGraphWithAlienFiles
MinimalGraphWithAlien files enumeration mode (mainly used for JS-based repos) was explicitly excluding directories from the fingerprint. This was causing an underbuild when a new directory (with files inside) was added without changing anything else. The canonical scenario for this behavior is relying on the TS compiler to glob all *.ts sources under the project root.
This PR enhances the dir enumeration fingerprint under MinimalGraphWithAlien files to include directories that were not produced by the build.
@<Kristijan Simic> , this PR also addresses a couple of edge cases in the managed resolver that the detours implementation should be already doing. After this PR gets in I'll try switching to the detours resolution and check against sp-client.

Related work items: #1783090