Merged PR 580196: Extract configuration and fix machine name propagation for copies
This PR moves the logic that copies configuration settings from `DistributedContentSettings` into domain objects from two places into one. Now, all the config classes have a helpers `FromDistributedContentSettings`.

This is done to make the code less error prone, because it should be much easier to spot that the configuration propagation was not correct after a new configuration was added.

Second, the machine name propagation is fixed and moved to `GrpcCopyClient` + added a test that shows the stuff works.