Merged PR 580701: Fix flakiness of PutAcrossRunsPurgesOldest test
This PR fixes the flakiness of `PutAcrossRunsPurgesOldest` test.

The test relies on memory clock to adjust the current time to separate the last access time for two hashes. Unfortunately, the test was moving the clock forward beyond the system's current time and that was causing the issue during `ContentDirectory` deserialization. The deserialization logic has a special check to adjust the last access time to the current time if the last access time in the file is beyond the current time.

And because `ContentDirectory` is not using the same memory clock, the last access time was adjusted during deserialization to `UtcNow`. And because deserialization logic is super fast in some cases it was possible to create the same last access time causing this test to fail.