Merged PR 581132: Ensures that output directories are tracked on output materialization
In the past because output directories are not tracked during output materialization, the journal processing has to do extra enumeration to verify that the output directories do not change its membership. At that time, we only tracked if the directories are produced by pips, but not when they come from cache.

This change adds unit tests to guarantee that the journal processing doesn't do any extra enumeration if no membership change.

Related work items: #1782665