Merged PR 582098: Add extra redis configuration settings and unified error handling logic
This PR propagates extra configuration that can be helpful for us for fine tuning the behavior (for instance in various github issues, folks suggesting setting KeepAlive, and @Julian discovered that setting `ConfigCheck` interval can be helpful as well.

This PR also changes `RedisConnectionMultiplexer` to use the same logic for two key operations: `ExecuteBatchOperationAsync` and `ExecuteDatabaseOperationAsync`.