Merged PR 582467: [JavaScript] Support grouping script commands
This PR adds the ability to group JavaScript script commands so BuildXL can treat them as a single pip. This is useful to ease the onboarding process. The typical case applies to the pre/post scripts associated to the standard scripts 'build', 'test', etc. E.g. prebuild, build and postbuild scripts can be just one pip. The main advantage is not having to deal with double writes between these scripts.
@<Danny van Velzen>  , I'm adding this for the Lage case as well, as discussed.