Merged PR 582160: Make enumeration fingerprint for MinimalWithAlienFiles mode more stable by preserving created directories on cache replay
Consider the following scenario:
* Build1
PipA creates dir D under dir Root
PipB enumerates Root. (D is identified as a created directory and therefore not added to Root's fingerprint)

* Build2
Shared opaque scrubbing happens. Directory D is deleted
PipA is replayed from the cache. D is recreated
PipB enumerates Root. The fact that D is a created directory needs to be known to get the same fingerprint as in Build1.

This PR deals with the case of Build2 by including the set of created directories into the cache. This still only happens when AllowedUndeclaredReads is on, which is what triggers this particular enumeration mode