Merged PR 582642: Track hashes in LLS-based operations in MDM
Currently we only track the total number of hashes, but not the number of hashes per each operation.

The current scheme can be misleading in case of mass re-images, when the number of remove events is quite small, but the number of hashes affected is huge.

This PR starts tracking the hashes per add/remove/touch and not only the number of affected hashes overall.