Merged PR 582879: Fail test gracefully if a contract is violated
Currently, if a contract is violated, the test will be terminated in 5 seconds.

The logic is not perfect, because, for instance, for .net core builds it makes the analysis of failure close to impossible because the test just crashes without writing its output.

This PR subscribes to contract violation even before running a test and then fails the test gracefully if any contract violations occurred during the test execution.

The logic is not 100% correct, because the tests can be run in parallel and it means that more than one test may fail because of a single contract violation. But because we don't expect too many tests to be affected by this, I think this logic is far better than the old one.