Merged PR 583438: Scrubber should not delete parents of temp directories
Pip specifies Out\D1\D2 as a temp directory. Pip probes Out\D1 during execution. After execution, Bxl's temp cleaner deletes Out\D1\D2, but (of course) not Out\D1. In the next build, the scrubber deletes Out\D1 because it is not in the graph, and is not an output directory. Unfortunately, the probe of Out\D1 is recorded in incremental scheduling state, and thus the deletion makes the pip dirty.

No-op builds can go from 1s to 5-6s because of this.

Related work items: #1790710