Merged PR 583688: Deal with absent probes containing reparse points that are later produced
Consider the following scenario:
1. File dir/to/file exists on disk, dir/junction does not
2. A pip probes dir/junction/file and then creates junction dir/junction -> dir/to

Even when full reparse point resolution is on, dir/junction/file is not resolved because when that access happens the path does not exist. However, observed input processor classifies the probe as present since after the pip is done, the path is there. Since dir/junction/file ends up in the path set and it is an unresolved path. on cache lookup this produces an inconsistency and the probe is absent, introducing a cache miss.

This PR addresses this problem by adding special handling for absent probes that later (in the scope of the same pip) becomes present, making sure they are properly resolved.

@<Kristijan Simic> , the plan is to keep ReparsePointResolver just to do GetFinalPathW with caching, but get rid of all the other functions there once we decide we can get rid of the managed resolution flag (which I think can happen next week, everything seems to going smoothly there)