Merged PR 584714: Add ability to have a write-only remote cache
Because of security concerns, we have determined that we want to be able to not read from the remote at all. For non-publishing builds, it's already doable with the `UseLocalOnly` flag. For publishing builds, we want to be able to specify that the cache is write-only.

For simplicity, rather than checking if the remote is write-only every time we want to do a read operation on the remote, I decided to solve this by passing in a different read-only cache altogether to the vertical cache aggregator. If the vertical aggregator receives a `remoteReadCache` different than null, it will use that to create read-only sessions instead of using the normal `remoteCache`.

In other words, when `IsWriteOnly` is true, `VerticalCacheAggregator` will now receive:
- `LocalCache : ServiceClientCache`
- `RemoteCache : BuildCache`
- `RemoteReadCache : MemCache` (In-memory cache which will remain empty because it is read-only)

When the flag is not specified, `RemoteReadCache` will be null and `VerticalCacheAggregator` will continue using the regular `RemoteCache` to create read-only sessions.

Related work items: #1791307