Merged PR 584916: Generate net5.0 qualifier by NugetSpecGenerator
This is the first out of a few PRs required to add support for .net 5 that was released a few days ago.

This PR adds 'net5.0' qualifier for every generated nuget package's spec the same way it adds 'netcoreapp3.1' and it should go first. This is important, because once the PR with an actual .net 5 support is published, in order for it to succeed, it needs to rely on the current LKG to generate the right qualifiers for all nuget packages during the build.

This PR is low risk.