Merged PR 584959: Update csc to 3.7.0 and fix nullability warnings
This is the second PR that is required to make the final PR for adding .net5 support simpler.

The PR updates the compiler to a 3.7 version (this version does not support C# 9 yet), and fixes a bunch of nullability warnings that appear after the upgrade to 3.7 and some of them will be visible only when targeting .net 5.

This is another low risk PR.