Merged PR 585237: Add support for NET 5
This PR adds support for .NET 5.

The only current limitation is that QTest is disabled in .NET 5 because it was failing with some "can't find the right assembly" issue that I'm going to investigate.

Also, the default qualifier stays the same and I need some help testing the changes on the mac and on Linux.

This PR also allow us using C# 9 features like records **for all target frameworks** and the following code compiles just fine:

```csharp
public record X
    {
        public int Y { get; init; }
    }
```

Keep in mind, that you need the most recently released VS in order to see intellisense for this code.