Merged PR 586446: Disable reparse point post processing on Mac by default
The equivalent of GetFinalPathByName is not implemented for Mac. But moreover, some probes operations are explicitly ignored for the Mac case (MacLookUp in particular) so even though the mac equivalent of the resolution function can be implemented, post-processing is not actually solving a problem for Mac.
If we happen to see other Mac probes in need of resolution we can considering implementing the missing bits. Disabling this feature by default will make the Mac validation green again.

Related work items: #1795101