Merged PR 586213: [JavaScript] Add support for customizing scheduling to enable QTest integration
This PR adds the ability to configure any JS-based resolver to execute a custom scheduling function defined in DScript. This function can decide how to schedule a pip representing a JavaScript project in its own way.
The frontend provides the discovered project information to the callback in order to help make decisions about how scheduling should happen.
This enables the integration with QTest, where the custom scheduling function can decide to schedule a QTest instead of a regular test.

Related work items: #1781495