Merged PR 586754: Allow MultiplexedStore to copy files between drives when PlaceFile with Hardlink-only is specified
Currently, we will fail if the destination file is not in the same drive as the one where we actually have the content.

Example:
1. Cache has ContentA on drive X.
2. `PlaceFile` with Hardlink and destination path=[D:\...]
3. FileSystem[D] fails with ContentNotFound
4. FileSystem[X] fails with Hardlink failed because they are in different drives.
5. Distributed store decides to do a `RemoteCopy` (?!) to put the content on drive D.
6. Reattempts the place file, at which point we create a hardlink from the cache in drive D

This enables us to copy from cache in `X:\` to the cache in `D:\` so that it can perform a hardlink.

I realize that this does not cover the bulk implementation of `PlaceFileAsync`, but this is only happening with QuickBuild and they don't use the bulk version, so I don't think it's worthwhile to do this change for the bulk version, since it would be significantly more difficult to do so and we won't see any benefits.

Testing: added a test that previously failed and now passes.

Related work items: #1795932