Merged PR 587459: Mitigate flakiness of BaselineTests.RetryPipOnHighMemoryUsage
Change the test to not rely on a cancellation to finish -- this was creating race conditions inhibiting the behavior to be tested.

Instead of launching processes that block and terminating the run via cancellation, schedule pips that wait for a file to exist, and create the file from the test when we want to finish. This still has potential for flakiness because we simulate high pressure for an arbitrary number of cycles and theoretically we could signal for termination before the scheduler has time to remove and cancel the pip, but it is more stable than the other race condition.

Related work items: #1794575