Merged PR 587738: Fail reconciliation if SendEventsAsync fails
This PR changes the logic to fail reconciliation operation if the events were not sent successfully to Event Hub.

This is achieved by passing a special configuration flag to `ContentLocationEventStore` to throw an error if `SendEventsAsync` fails. In this case, the nagle queue switches to an error state and the error is observed when the queue is disposed.