Merged PR 582877: Unify and disambiguate identifier(s) in BXL and ADO - Phase 2
**Problem**
Like the `BlobIdentifier` the DedupIdentifer currently lives dual lives - one in BXL and ADO. The aim is remove this fork and have one unified place for such identifiers and the corresponding logic.

**Changes**
This change makes BXL's `DedupIdentifier`, and its derivations (`NodeDedupIdentifier` and `ChunkDedupIdentifier`) to be wire-to-wire compatible with the one in ADO.
Add the ability to get the content hasher, given a `NodeAlgorithmId`.

**Future Changes**
Update DropDaemon -> to use the correct BlobIdentifier.
Update DropClient.Slave -> to use the correct BlobIdentifier.
Update BuildCache server -> to use the correct BlobIdentifier.
Get rid of VstsBlobIdentifer converter from BXL repo since we won't be needing it anymore.
Make ADO Repo happy with the BXL BlobIdentifier.
Make ADO Repo happy with the DedupIdentifier(s).

Related work items: #1789340